move ApplicationModel

+++ benchmark-base/src/main/java/benchmark/rpc/route/RouteService.java
@@ -1,13 +1,13 @@
 package benchmark.rpc.route;
 
+import benchmark.bean.User;
+import benchmark.service.UserService;
+import benchmark.service.UserServiceServerImpl;
+
 import static benchmark.service.ServiceRegister.CREATE_USER;
 import static benchmark.service.ServiceRegister.EXIST_USER;
 import static benchmark.service.ServiceRegister.GET_USER;
 import static benchmark.service.ServiceRegister.LIST_USER;
-
-import benchmark.bean.User;
-import benchmark.service.UserService;
-import benchmark.service.UserServiceServerImpl;
 
 public class RouteService {
 
@@ -23,7 +23,7 @@ public class RouteService {
 			return userService.createUser((User) params[0]);
 
 		case GET_USER:
-			return userService.getUser((Long) params[0]);
+			return userService.getUser((Integer) params[0]);
 
 		case LIST_USER:
 			return userService.listUser((Integer) params[0]);
